% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eye.approximate.ReL1}
\alias{eye.approximate.ReL1}
\title{Calculates various approximations for the real part of the leading eigenvalue
of the community matrix M.}
\usage{
eye.approximate.ReL1(M, calculate_eigenvalues = TRUE)
}
\arguments{
\item{M}{The community matrix. Note that for the Allesina et al. (eyeball) approximation, M should contain (almost) no
positive coefficients in the upper triangular part and (almost) no negative coefficients in the lower-triangular
part}

\item{calculate_eigenvalues}{(TRUE by default) calculate the actual eigenvalues of M and return them}
}
\value{
A list composed of:
\describe{
\item{$M}{The community matrix}
\item{$M.eigenvalues}{The eigenvalues of M (if calculate_eigenvalues = TRUE)}
\item{$ReL1.observed }{The actual real part of the rightmost eigenvalue of M}
\item{$ReL1.May }{The prediction according to May's approximation}
\item{$ReL1.TangEtAl}{The prediction according to Tang et al. approximation}
\item{$ReL1.eyeball}{The prediction according to the "eyeball" approach in Allesina et al. }
\item{$May.stats}{List containing the relevant statistics for computing May's approximation}
\item{$TangEtAl.stats}{List containing the relevant statistics for computing the approximation of Tang et al.}
\item{$eyeball.stats}{List containing the relevant statistics for computing the eyeball approximation of Allesina et al.}
}
}
\description{
It returns the expectation for the real part of the rightmost eigenvalue of M
according to May (Nature 1972), Tang et al (Ecology Letters 2014), and
Allesina et al. (Nature Communications 2015) derivations.
}
\examples{
M <- eye.parameterize.M(eye.foodweb.cascade())
eye.approximate.ReL1(M)
}

