% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eye.buildfoodweb.and.parameterize.M}
\alias{eye.buildfoodweb.and.parameterize.M}
\title{Builds a community matrix M based on a food web model, and a parameterization for the bivariate distribution}
\usage{
eye.buildfoodweb.and.parameterize.M(foodweb_model = "Cascade", size = 100,
  connectance = 0.1, distribution_pairs = "Normal", mux = -3, muy = 1.5,
  sigmax = 1, sigmay = 0.75, rhoxy = -2/3, mu_diagonal = 0,
  sigma_diagonal = 0)
}
\arguments{
\item{foodweb_model}{A food web model. Possible options are "Random", "Cascade", "Niche" or a file name containing an adjacency matrix}

\item{size}{The number of species in the foodweb}

\item{connectance}{The connectance of the food web}

\item{distribution_pairs}{A bivariate distribution for the pairs (possibilities are "Normal", "FourCorner", or an empirical distribution, represented by a Nx2 matrix of pairs)}

\item{mux}{Mean of the negative interaction strengths}

\item{muy}{Mean of the positive interaction strengths}

\item{sigmax}{Standard deviation of the negative interaction strengths}

\item{sigmay}{Standard deviation of the positive interaction strengths}

\item{rhoxy}{Correlation between the interaction strengths}

\item{mu_diagonal}{(0 by default) mean of the diagonal elements}

\item{sigma_diagonal}{(0 by default) standard deviation for the diagonal elements}
}
\value{
The community matrix M
}
\description{
Builds a community matrix M based on a food web model, and a parameterization for the bivariate distribution
}
\examples{
M <- eye.buildfoodweb.and.parameterize.M(foodweb_model = "Cascade", size = 100, connectance = 0.2,
     distribution_pairs = "Normal", mux = -1, muy = 1, sigmax = 1, sigmay = 1, rhoxy = -0.2,
     mu_diagonal = -3, sigma_diagonal = 0)
}

