% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eye.foodweb.file}
\alias{eye.foodweb.file}
\title{Build a food web reading a file}
\usage{
eye.foodweb.file(filename)
}
\arguments{
\item{filename}{A text file containing the adjacency matrix (binary, rows = resources, cols = consumers) of a food web
The food web is a list containing:}
}
\value{
The food web is a list containing:

\describe{
\item{$links}{The trophic links from resource to consumer (L x 2 matrix)}
\item{$size}{The number of species}
\item{$L}{The number of links}
\item{$connectance}{The connectance L / (size choose 2)}
\item{$model}{A code for the model used to generate the food web}
\item{$cycles}{Boolean, whether cycles are present}
}
All the self-loops and double arrows (a eats b, b eats a) are removed. If the food web contains
no cycles, then it is sorted that all the coefficients would be in the upper-triangular part
of the corresponding adjacency matrix.
}
\description{
Read the binary adjacency matrix contained in the file.
Remove self-loops and double edges
}

