% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eye.pairs.from.empirical}
\alias{eye.pairs.from.empirical}
\title{Sample pairs of interactions from an empirical distribution}
\usage{
eye.pairs.from.empirical(num_pairs = 10, empirical_distribution = NULL)
}
\arguments{
\item{num_pairs}{The number of pairs to sample}

\item{empirical_distribution}{A Nx2 matrix with the effects of consumers on resources
                               in the first column and those of resources on consumers in the second.}
}
\value{
A num_pairs x 2 matrix of interaction strengths sampled from the distribution
}
\description{
Sample pairs of interactions from an empirical distribution
}
\examples{
mypairs <- eye.pairs.from.empirical(100, matrix(runif(20), 10, 2))
}

