% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eye.pairs.from.normal}
\alias{eye.pairs.from.normal}
\title{Sample pairs of interactions from a bivariate normal distribution}
\usage{
eye.pairs.from.normal(num_pairs = 10, mux = -1, muy = 0.5, sigmax = 1/4,
  sigmay = 1/4, rhoxy = -2/3)
}
\arguments{
\item{num_pairs}{The number of pairs to sample}

\item{mux}{The desired mean for the first marginal distribution}

\item{muy}{The desired mean for the second marginal distribution}

\item{sigmax}{The desired standard deviation for the first marginal distribution}

\item{sigmay}{The desired standard deviation for the second marginal distribution}

\item{rhoxy}{The desired correlation between the pairs of interaction strengths}
}
\value{
A num_pairs x 2 matrix of interaction strengths sampled from the distribution
}
\description{
Sample pairs of interactions from a bivariate normal distribution
}
\examples{
mypairs <- eye.pairs.from.normal() # default values
mypairs <- eye.pairs.from.normal(450)
mypairs <- eye.pairs.from.normal(num_pairs = 450)
mypairs <- eye.pairs.from.normal(num_pairs = 450, mux = 0, muy = 0,
           sigmax = 1, sigmay = 1, rhoxy = 0) # as for May's matrices
}

