% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eye.parameterize.M}
\alias{eye.parameterize.M}
\title{Builds a community matrix M based on a food web, and a parameterization for the bivariate distribution}
\usage{
eye.parameterize.M(FW, distribution_pairs = "Normal", mux = -1, muy = 0.5,
  sigmax = 1/4, sigmay = 1/4, rhoxy = -2/3, mu_diagonal = 0,
  sigma_diagonal = 0)
}
\arguments{
\item{FW}{A food web object, produced by the method eye.foodweb.cascade or similar}

\item{distribution_pairs}{A bivariate distribution for the pairs (possibilities are "Normal", "FourCorner", or an empirical distribution, represented by a Nx2 matrix of pairs)}

\item{mux}{Mean of the negative interaction strengths}

\item{muy}{Mean of the positive interaction strengths}

\item{sigmax}{Standard deviation of the negative interaction strengths}

\item{sigmay}{Standard deviation of the positive interaction strengths}

\item{rhoxy}{Correlation between the interaction strengths}

\item{mu_diagonal}{(0 by default) mean of the diagonal elements}

\item{sigma_diagonal}{(0 by default) standard deviation for the diagonal elements}
}
\value{
The community matrix M
}
\description{
Builds a community matrix M based on a food web, and a parameterization for the bivariate distribution
}
\examples{
M <- eye.parameterize.M(eye.foodweb.cascade())
M <- eye.parameterize.M(eye.foodweb.niche(size = 30, connectance = 0.2),
     distribution_pairs = "FourCorner", mux = -1, muy = 1, sigmax = 2, sigmay = 1, rhoxy = -0.5)
}

